update method to get, update, delete for users only

+++ src/TaskGateway.php
@@ -32,7 +32,7 @@ class TaskGateway
         $data = [];
 
         // return fetched data all data
-        // return $stmt->fetchAll(PDO::FETCH_ASSOC);
+        // return $stmt->fetchAll(PDO::FETCH_ASSOC); 
         // fetch rows
         while ($row = $stmt->fetch(PDO::FETCH_ASSOC)) {
             // convert to bool true or false
@@ -44,17 +44,19 @@ class TaskGateway
     }
 
     // get single record
-    public function get(string $id): array | false
+    public function getForUser(int $user_id, string $id): array | false
     {
         $sql = "SELECT *
                     FROM task
-                    WHERE id = :id";
+                    WHERE id = :id
+                    AND user_id = :user_id";
 
         // to prevent sql injection we make statement
         $stmt = $this->conn->prepare($sql);
 
         // bind param to id
         $stmt->bindValue(":id", $id, PDO::PARAM_INT);
+        $stmt->bindValue(":user_id", $user_id, PDO::PARAM_INT);
 
         // execute statement
         $stmt->execute();
@@ -70,10 +72,10 @@ class TaskGateway
     }
 
     // create task
-    public function create(array $data): string
+    public function createForUser(int $user_id, array $data): string
     {
-        $sql = "INSERT INTO task (name, priority, is_completed) 
-                    VALUES (:name, :priority, :is_completed)";
+        $sql = "INSERT INTO task (name, priority, is_completed, user_id) 
+                    VALUES (:name, :priority, :is_completed, :user_id)";
 
         $stmt = $this->conn->prepare($sql);
 
@@ -92,12 +94,14 @@ class TaskGateway
 
         $stmt->bindValue(":is_completed", $data["is_completed"] ?? false, PDO::PARAM_BOOL);
 
+        $stmt->bindValue(":user_id", $user_id, PDO::PARAM_INT);
+
         $stmt->execute();
 
         return $this->conn->lastInsertId();
     }
 
-    public function update(string $id, array $data): int
+    public function updateForUser(int $user_id, string $id, array $data): int
     {
         $fields = [];
 
@@ -136,11 +140,13 @@ class TaskGateway
         // create sql
         $sql = "UPDATE task"
                 . " SET " . implode(", ", $sets)
-                . " WHERE id = :id";
+                . " WHERE id = :id"
+                . " AND user_id = :user_id";
         
         $stmt = $this->conn->prepare($sql);
 
         $stmt-> bindValue(":id", $id, PDO::PARAM_INT);
+        $stmt->bindValue(":user_id", $user_id, PDO::PARAM_INT);
 
         // bind param for each fields in array
         foreach ($fields as $name => $values) {
@@ -155,14 +161,16 @@ class TaskGateway
         }        
     }
 
-    public function delete(string $id): int
+    public function deleteForUser(int $user_id, string $id): int
     {
         $sql = "DELETE FROM task
-                    WHERE id = :id";
+                    WHERE id = :id
+                    AND user_id = :user_id";
         
         $stmt = $this->conn->prepare($sql);
 
         $stmt->bindValue(":id", $id, PDO::PARAM_INT);
+        $stmt->bindValue(":user:id", $user_id, PDO::PARAM_INT);
 
         $stmt->execute();
